Fix page routing
