+++ packages/modules/web_themes/standard/source/src/router/routes.ts
@@ -4,13 +4,13 @@ const routes: RouteRecordRaw[] = [
   {
     path: '/info',
     component: () => import('layouts/MainLayout.vue'),
-    children: [{ path: '', component: () => import('pages/IndexPage.vue') }],
+    children: [{ path: '', component: () => import('pages/InfoPage.vue') }],
   },
 
   {
     path: '/',
     component: () => import('layouts/MainLayout.vue'),
-    children: [{ path: '', component: () => import('pages/TestPage.vue') }],
+    children: [{ path: '', component: () => import('pages/IndexPage.vue') }],
   },
 
   // Always leave this as last one,
